# python-POC/test-server/go-server-server-generated/api/swagger.yaml
openapi: 3.0.1
info:
  title: Danish Data Portal - HPC API
  description: A Web API for HPC centers to communicate with the Danish Data Portal
    to manage access and data transfers
  version: v1
servers:
- url: /
security:
- ApiKey: []
paths:
...
  /messages:
    get:
      tags:
      - Messages
      summary: Fetches all unprocessed messages adressed to the caller
      description: "The data field relates to the messageType and is a JSON string\
        \ with the relevant data. Below the data field is described for each messageType.\r\
        \n            \r\nmessageType = CreateProject:\r\n            \r\n    {\r\n\
        \       \"ProjectNo\": \"712456\",\r\n    }\r\n            \r\nmessageType\
        \ = DeleteProject:\r\n            \r\n    {\r\n       \"ProjectNo\": \"712456\"\
        ,\r\n    }\r\n            \r\nmessageType = CreateProjectAccess:\r\n     \
        \       \r\n    {\r\n       \"ProjectNo\": \"712456\",\r\n       \"AccessIdentifier\"\
        : \"3cb99f77-5c83-41fa-ac15-06f90890c66e\",\r\n    }\r\n            \r\nmessageType\
        \ = DeleteProjectAccess:\r\n            \r\n    {\r\n       \"AccessIdentifier\"\
        : \"3cb99f77-5c83-41fa-ac15-06f90890c66e\"\r\n    }\r\n            \r\nmessageType\
        \ = DisableProjectAccess:\r\n            \r\n    {\r\n       \"AccessIdentifier\"\
        : \"3cb99f77-5c83-41fa-ac15-06f90890c66e\"\r\n    }\r\n            \r\nmessageType\
        \ = EnableProjectAccess:\r\n            \r\n    {\r\n       \"AccessIdentifier\"\
        : \"3cb99f77-5c83-41fa-ac15-06f90890c66e\"\r\n    }\r\n            \r\nmessageType\
        \ = ResetPassword:\r\n            \r\n    {\r\n       \"AccessIdentifier\"\
        : \"3cb99f77-5c83-41fa-ac15-06f90890c66e\"\r\n    }\r\n            \r\nmessageType\
        \ = DataDeliveryReady:\r\n            \r\n    {\r\n       \"DataDeliveryId\"\
        :\"2cb99f77-5c83-41fa-ac15-06f90890c66e\",\r\n       \"ProjectNumber\":712456,\r\
        \n       \"Files\": [\r\n          {\r\n              \"FileId\":\"fa5a6bff-2d35-40e4-86d9-e90945732720\"\
        ,\r\n              \"FileSize\":22,\r\n              \"FileChecksum\":\"5f7948af5e9db4fc545b49a09f3c37d1\"\
        \r\n          }\r\n       ]\r\n    }\r\n            \r\nmessageType = DeleteDataFile:\r\
        \n            \r\n    {\r\n       \"FileId\":\"fa5a6bff-2d35-40e4-86d9-e90945732720\"\
        \r\n    }\r\n            \r\n            \r\nmessageType = ReturnDataFile:\r\
        \n            \r\n    {\r\n       \"FileId\":\"fa5a6bff-2d35-40e4-86d9-e90945732720\"\
        \r\n    }"
      responses:
        "200":
          description: Returns a list of messages
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageDTO'
                x-content-type: text/plain
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageDTO'
        "401":
          description: Unauthorized - wrong or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
    patch:
      tags:
      - Messages
      summary: Marks all specified messages as processed.
      requestBody:
        description: List of GUIDs of messages to mark as processed
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                type: string
                format: uuid
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uuid
          text/json:
            schema:
              type: array
              items:
                type: string
                format: uuid
          application/*+json:
            schema:
              type: array
              items:
                type: string
                format: uuid
      responses:
        "200":
          description: Messages marked as processed
        "400":
          description: One or more message was already marked as processed
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "401":
          description: Unauthorized - wrong or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: One or more message does not belong to the calling HPC center
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: One or more message does not exist
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
components:
  schemas:
    ErrorDetails:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
          nullable: true
      additionalProperties: false
    HPCMessageTypeEnum:
      type: string
      enum:
      - CreateProject
      - DeleteProject
      - CreateProjectAccess
      - DeleteProjectAccess
      - DisableProjectAccess
      - EnableProjectAccess
      - ResetPassword
      - DataDeliveryReady
      - DeleteDataFile
      - ReturnDataFile
    MessageDTO:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        messageType:
          $ref: '#/components/schemas/HPCMessageTypeEnum'
        timestamp:
          type: string
          format: date-time
        data:
          type: string
          nullable: true
      additionalProperties: false
      example:
        messageType: CreateProject
        data: data
        messageId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        timestamp: 2000-01-23T04:56:07.000+00:00
...
  securitySchemes:
    ApiKey:
      type: apiKey
      description: "API key needed to access the endpoints. Example: 'ApiKey: YOUR_API_KEY'"
      name: X-API-Key
      in: header